Code/Data_Extraction/requirements.txt
functions-framework==3.*
pandas
pyarrow
requests
google-cloud-storage